python_sdk_samples/pyproject.toml
[project]
name = "python-sdk-samples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cadence-python-client[openai]>=0.2.1",
    "ruff>=0.15.10",
]
